Expiration Date Update
Updates to the expiration dates to respect the 2038 bug to ensure
optimal browser compatibility.  With this fix, a cookie can/will not
expire beyond January 1st, 2038 - which is a limitation of
browsers/computers at this time.  (technically the last available date
is January 19th, however this script sets it to January 1st).

Reference:
https://en.wikipedia.org/wiki/Year_2038_problem